EditComponentWizard rewired to fix multi ppl edit bug
Originally the 'component' viewState was added to the wizard at the
time when the wizard was instantiated. This is bad because on the
wizard, self.states is a class attribute.
This fix, creates a state for each existing active component when
the wizard is __setup__ in the trytond engine. When a component
is selected, the appropriate state is shown based on the type
of the component. States are named according to the shortname
of the component shown in the first column.

And bump version number for this bugfix.